backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]
sqlalchemy==2.0.*
alembic
python-multipart
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]